requirements.txt
PyQt6
parse-torrent-name
thefuzz
python-levenshtein
ffmpeg-python